Refactor & Remove Dead Code
