+++ composer.json
@@ -41,8 +41,7 @@
     },
     "autoload": {
         "psr-4": {
-            "DDTrace\\": "./src/api/",
-            "DDTrace\\": "./src/DDTrace/"
+            "DDTrace\\": "./src/api/"
         },
         "files": [
             "./src/api/ComposerBootstrap.php"
@@ -51,8 +50,7 @@
     "autoload-dev": {
         "psr-4": {
             "DDTrace\\Tests\\Api\\": "./tests_api/",
-            "DDTrace\\Tests\\": "./tests/",
-            "DDTrace\\Benchmarks\\": "./tests/"
+            "DDTrace\\Tests\\": "./tests/"
         }
     },
     "minimum-stability": "dev",
