+++ tests/Benchmarks/Integrations/SymfonyBench.php
@@ -2,16 +2,44 @@
 
 declare(strict_types=1);
 
-namespace DDTrace\Benchmarks;
+namespace Benchmarks\Integrations;
 
-use DDTrace\Tests\Common\TracerTestTrait;
-use DDTrace\Tests\Common\Utils;
 use DDTrace\Tests\Common\WebFrameworkTestCase;
 use DDTrace\Tests\Frameworks\Util\Request\GetSpec;
 
 class SymfonyBench extends WebFrameworkTestCase
 {
-    use TracerTestTrait;
+    /**
+     * @BeforeMethods("disableSymfonyTracing")
+     * @AfterMethods("afterMethod")
+     * @Revs(10)
+     * @Iterations(10)
+     * @OutputTimeUnit("microseconds")
+     * @RetryThreshold(10.0)
+     */
+    public function benchSymfonyBaseline()
+    {
+        $this->call(GetSpec::create(
+            'A simple GET request with a view',
+            '/simple_view?key=value&pwd=should_redact'
+        ));
+    }
+
+    /**
+     * @BeforeMethods("enableSymfonyTracing")
+     * @AfterMethods("afterMethod")
+     * @Revs(10)
+     * @Iterations(10)
+     * @OutputTimeUnit("microseconds")
+     * @RetryThreshold(10.0)
+     */
+    public function benchSymfonyOverhead()
+    {
+        $this->call(GetSpec::create(
+            'A simple GET request with a view',
+            '/simple_view?key=value&pwd=should_redact'
+        ));
+    }
 
     public static function getAppIndexScript()
     {
@@ -35,37 +63,5 @@ class SymfonyBench extends WebFrameworkTestCase
         $this->setUpWebServer([
             'DD_TRACE_ENABLED' => 1,
         ]);
-    }
-
-    /**
-     * @BeforeMethods("disableSymfonyTracing")
-     * @AfterMethods("afterMethod")
-     * @Revs(1)
-     * @Iterations(10)
-     * @OutputTimeUnit("microseconds")
-     * @RetryThreshold(10.0)
-     */
-    public function benchSymfonyBaseline()
-    {
-        Utils::call(GetSpec::create(
-            'A simple GET request with a view',
-            '/simple_view?key=value&pwd=should_redact'
-        ));
-    }
-
-    /**
-     * @BeforeMethods("enableSymfonyTracing")
-     * @AfterMethods("afterMethod")
-     * @Revs(1)
-     * @Iterations(10)
-     * @OutputTimeUnit("microseconds")
-     * @RetryThreshold(10.0)
-     */
-    public function benchSymfonyOverhead()
-    {
-        Utils::call(GetSpec::create(
-            'A simple GET request with a view',
-            '/simple_view?key=value&pwd=should_redact'
-        ));
     }
 }
